+++ ace0.5/offer.h
@@ -10,24 +10,24 @@ class offer
 {
 public:
 	offer(void);
-	offer(double price, int count): _price(price), _count(count)
+	offer(double price, double count): _price(price), _count(count)
 	{
 	}
 	double get_price();
-	int get_count();
+	double get_count();
 	
 	void set_count(double count);
 
 	int getid();
 
 private:
-	int _count;
+	double _count;
 	double _price;
 };
 
 int get_random(map<int, offer> good);
 
-int get_count(int firmid, map<int, offer> good);
+double get_count(int firmid, map<int, offer> good);
 
 bool can_buy(double money, map<int, offer> good);
